requirements.txt
flask
requests
python-dotenv
google-generativeai
psycopg2-binary
gunicorn
APScheduler
datetime
re